pyproject.toml
[build-system]
requires    = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name            = "a2a-server"
version         = "0.1.0"
description     = "In‑memory FSM for the A2A Protocol with Pydantic v2 models"
readme          = "README.md"
requires-python = ">=3.11"

# runtime dependencies
dependencies = [
  "a2a-json-rpc>=0.1.3",
  "asyncio>=3.4.3",
  "pydantic>=2.11.3",
  "fastapi>=0.115.12",
  "uvicorn>=0.34.1",
  "httpx>=0.28.1",
  "pyyaml>=6.0.2",
  "google-adk>=0.2.0",
  "litellm>=1.66.3",
]

# optional feature sets
[project.optional-dependencies]
dev = [
  "pytest>=7.0.0",
  "pytest-asyncio>=0.20.0"
]

# Tell setuptools to look in src/ for your a2a package
[tool.setuptools.packages.find]
where   = ["src"]
include = ["a2a_server*"]

[tool.pdm]
python_requires = ">=3.11"

# Convenient CLI shortcuts
[tool.pdm.scripts]
test            = "pytest"

[project.scripts]
a2a-server = "a2a_server.__main__:main"

# pytest settings so it finds your src/ layout automatically
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v"
asyncio_default_fixture_loop_scope = "function"
asyncio_mode = "strict"

filterwarnings = [
  # Protobuf “PyType_Spec” warnings
  "ignore:Type google\\._upb\\._message\\.MessageMapContainer uses PyType_Spec:DeprecationWarning",
  "ignore:Type google\\._upb\\._message\\.ScalarMapContainer uses PyType_Spec:DeprecationWarning",

  # dateutil UTC‐timestamp deprecation
  "ignore:datetime\\.datetime\\.utcfromtimestamp\\(\\) is deprecated:DeprecationWarning",

  # pkg_resources / namespace package deprecations
  "ignore:pkg_resources is deprecated as an API:DeprecationWarning",
  "ignore:Deprecated call to `pkg_resources\\.declare_namespace\\('google'\\)`:DeprecationWarning",
  "ignore:Deprecated call to `pkg_resources\\.declare_namespace\\('google\\.cloud'\\)`:DeprecationWarning",

  # (Optional) catch all other DeprecationWarnings
  "ignore::DeprecationWarning",
]
